requirements.txt
dj_database_url
django
django-admin-sortable2
django-tinymce
environs
pillow
requests